simplified and add minor fixes in ocr module
