--- booksbridge/src/components/Dropzone/Dropzone.js
@@ -1,102 +1,0 @@
-import React, { Component } from 'react';
-import './Dropzone.css';
-
-class Dropzone extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { hightlight: false, image: null };
-    this.fileInputRef = React.createRef();
-
-    this.openFileDialog = this.openFileDialog.bind(this);
-    this.onFilesAdded = this.onFilesAdded.bind(this);
-    this.onDragOver = this.onDragOver.bind(this);
-    this.onDragLeave = this.onDragLeave.bind(this);
-    this.onDrop = this.onDrop.bind(this);
-  }
-
-  openFileDialog() {
-    if (this.props.disabled) return;
-    this.fileInputRef.current.click();
-  }
-
-  onFilesAdded(event) {
-    if (this.props.disabled) return;
-
-    const { files } = event.target;
-
-    if (event.target.files && event.target.files[0]) {
-      let reader = new FileReader();
-      reader.onload = e => {
-        this.setState({ image: e.target.result });
-      };
-      reader.readAsDataURL(event.target.files[0]);
-    }
-
-    if (this.props.onFilesAdded) {
-      const array = this.fileListToArray(files);
-      this.props.onFilesAdded(array);
-    }
-  }
-
-  onDragOver(evt) {
-    evt.preventDefault();
-
-    if (this.props.disabled) return;
-
-    this.setState({ hightlight: true });
-  }
-
-  onDragLeave() {
-    this.setState({ hightlight: false });
-  }
-
-  onDrop(event) {
-    event.preventDefault();
-
-    if (this.props.disabled) return;
-
-    const { files } = event.dataTransfer;
-    if (this.props.onFilesAdded) {
-      const array = this.fileListToArray(files);
-      this.props.onFilesAdded(array);
-    }
-    this.setState({ hightlight: false });
-  }
-
-  fileListToArray(list) {
-    const array = [];
-    for (let i = 0; i < list.length; i++) {
-      array.push(list.item(i));
-    }
-    return array;
-  }
-
-  render() {
-    return (
-      <div className="dropzone">
-        <div
-          className={`Dropzone ${this.state.hightlight ? 'Highlight' : ''}`}
-          onDragOver={this.onDragOver}
-          onDragLeave={this.onDragLeave}
-          onDrop={this.onDrop}
-          onClick={this.openFileDialog}
-          style={{ cursor: this.props.disabled ? 'default' : 'pointer' }}
-        >
-          <input
-            ref={this.fileInputRef}
-            className="FileInput"
-            type="file"
-            accept=".jpg, .jpeg, .png"
-            multiple
-            onChange={this.onFilesAdded}
-          />
-          <img alt="upload" className="Icon" src="../../../upload.png" />
-          <span>Upload Files</span>
-        </div>
-        <img id="target" src={this.state.image} width={300} height={300} />
-      </div>
-    );
-  }
-}
-
-export default Dropzone;
