+++ booksbridge/src/store/reducers/book.js
@@ -5,7 +5,7 @@ const initialState = {
   count: 0,
   selectedBook: null,
   searchedBooks: [],
-  quote: 'empty',
+  quote: '',
 };
 
 const reducer = (state = initialState, action) => {
@@ -22,7 +22,6 @@ const reducer = (state = initialState, action) => {
     case actionTypes.EMPTY_SEARCHED_BOOKS:
       return { ...state, searchedBooks: [] };
     case actionTypes.RUN_OCR:
-      console.log(action.quote);
       return { ...state, quote: action.quote };
   }
